+++ lib/rules/no-invalid-position-declaration/index.mjs
@@ -1,7 +1,7 @@
 import { isAtRule, isRule } from '../../utils/typeGuards.mjs';
 import { isRegExp, isString } from '../../utils/validateTypes.mjs';
 import findNodeUpToRoot from '../../utils/findNodeUpToRoot.mjs';
-import isInInlineDocument from '../../utils/isInInlineDocument.mjs';
+import isInDocument from '../../utils/isInDocument.mjs';
 import isStandardSyntaxDeclaration from '../../utils/isStandardSyntaxDeclaration.mjs';
 import { nestingSupportedAtKeywords } from '../../reference/atKeywords.mjs';
 import optionsMatches from '../../utils/optionsMatches.mjs';
@@ -42,7 +42,7 @@ const rule = (primary, secondaryOptions) => {
 			isAtRule(node) && optionsMatches(secondaryOptions, 'ignoreAtRules', node.name);
 
 		root.walkDecls((decl) => {
-			if (isInInlineDocument(decl)) return;
+			if (isInDocument(decl, { inline: true })) return;
 
 			if (!isStandardSyntaxDeclaration(decl)) return;
