+++ lib/utils/__tests__/isInDocument.test.mjs
@@ -231,4 +231,75 @@ describe('isInDocument', () => {
 			expect(isInDocument(decl)).toBe(true);
 		});
 	});
+
+	describe('with postcss-html parser (inline)', () => {
+		it('returns false for declarations in HTML style tags', () => {
+			const document = postcssHtml.parse('<style>a { color: red; --custom: blue; }</style>');
+			const decls = [];
+
+			document.walkDecls((decl) => decls.push(decl));
+
+			expect(decls).toHaveLength(2);
+			expect(isInDocument(decls[0], { inline: true })).toBe(false);
+			expect(isInDocument(decls[1], { inline: true })).toBe(false);
+		});
+
+		it('returns false for deeply nested declarations in HTML', () => {
+			const document = postcssHtml.parse(`
+				<style>
+					@media (min-width: 600px) {
+						.container {
+							display: grid;
+							@supports (display: grid) {
+								grid-template-columns: 1fr 1fr;
+							}
+						}
+					}
+				</style>
+			`);
+			const decls = [];
+
+			document.walkDecls((decl) => decls.push(decl));
+
+			expect(decls).toHaveLength(2);
+			expect(isInDocument(decls[0], { inline: true })).toBe(false);
+			expect(isInDocument(decls[1], { inline: true })).toBe(false);
+		});
+
+		it('handles edge case where root.parent might be undefined', () => {
+			// Test cases where HTML parsers may not properly set parent relationships
+			// Some parsers create Document nodes but don't always link Root.parent to Document
+			const document = postcssHtml.parse('<style>a { color: red; }</style>');
+			const root = document.first;
+			const decl = root.first.first;
+
+			// Verify the expected AST structure: Document -> Root -> Rule -> Declaration
+			expect(document.type).toBe('document');
+			expect(root.type).toBe('root');
+
+			// In some HTML parsing scenarios, root.parent may be undefined
+			// despite the root being contained within a document
+			expect(root.parent).toBeUndefined();
+
+			// Our function should still correctly detect that the declaration
+			// is within a document context through alternative means
+			expect(isInDocument(decl, { inline: true })).toBe(false);
+		});
+
+		it('returns false for declarations in multiple HTML style sources', () => {
+			const document = postcssHtml.parse(`
+				<style>body { margin: 0; }</style>
+				<div style="color: red;">content</div>
+				<style>.class { padding: 10px; }</style>
+			`);
+			const decls = [];
+
+			document.walkDecls((decl) => decls.push(decl));
+
+			expect(decls).toHaveLength(3);
+			expect(isInDocument(decls[0], { inline: true })).toBe(false); // margin: 0
+			expect(isInDocument(decls[1], { inline: true })).toBe(true); // color: red
+			expect(isInDocument(decls[2], { inline: true })).toBe(false); // padding: 10px
+		});
+	});
 });
